% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.aplms.R
\name{AIC.aplms}
\alias{AIC.aplms}
\title{Akaike information criterion}
\usage{
\method{AIC}{aplms}(object, ...)
}
\arguments{
\item{object}{APLMS object.}

\item{...}{other arguments}
}
\value{
a numeric value of the corresponding AIC.
}
\description{
Print the AIC of the fitted APLMS model.
}
\examples{
\dontrun{AIC(object)}
}
