% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplms2.R
\name{aplms2}
\alias{aplms2}
\title{Fitting Additive partial linear models with symmetric errors}
\usage{
aplms2(
  formula,
  npc,
  basis,
  Knot,
  data,
  family = Normal(),
  p = 1,
  control = list(tol = 0.001, algorithm1 = c("backfitting"), algorithm2 = c("BFGS")),
  init,
  lam
)
}
\arguments{
\item{formula}{formula}

\item{npc}{non parametric component}

\item{basis}{basis}

\item{Knot}{knots}

\item{data}{data frame}

\item{family}{Symmetric error distribution}

\item{p}{autoregressive order of the error}

\item{control}{optimization rutine}

\item{init}{initiation}

\item{lam}{smoothing parameter vector}

\item{...}{other arguments}
}
\description{
Fitting Additive partial linear models with symmetric errors
}
\details{
\code{aplms2} is used to fit additive partial linear models with symmetric errors.
In this setup, the natural cubic splines or cubic P-splines.
}
\examples{
a <- 1

}
