% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculatef_init}
\alias{calculatef_init}
\title{Calculate the first aproximation of the gamma parameters.}
\usage{
calculatef_init(k, y, N_i)
}
\arguments{
\item{k}{number of non parametric components.}

\item{y}{observed values of the response variable.}

\item{N_i}{function matrices.}
}
\description{
Calculate the first aproximation of the gamma parameters.
}
