% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_FF}
\alias{calculateLL_FF}
\title{Calculate the second derivative with respect to the gamma estimations}
\usage{
calculateLL_FF(phi, Dd, Dv, AN, k, const2)
}
\arguments{
\item{phi}{vector of the estimated phi parameters.}

\item{Dd}{diagonal of the d1, ..., dn vectors}

\item{Dv}{diagonal of the v1, ..., vn vectors}

\item{AN}{vector of the function matrices multiplied by the matrix A}

\item{k}{number of non parametric components.}

\item{const2}{constant}
}
\description{
Calculate the second derivative with respect to the gamma estimations
}
