% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_FF_Phi}
\alias{calculateLL_FF_Phi}
\title{Calculate the second derivative with respect to the gamma estimations and phi estimations}
\usage{
calculateLL_FF_Phi(phi, AN, Dd, Dv, A, error_hat)
}
\arguments{
\item{phi}{vector of the estimated phi parameters.}

\item{AN}{vector of the function matrices multiplied by the matrix A}

\item{Dd}{diagonal of the d1, ..., dn vectors}

\item{Dv}{diagonal of the v1, ..., vn vectors}

\item{A}{A matrix}

\item{error_hat}{residual of the model}
}
\value{
second derivative of the log function with respect to tha gamma estimations and phi estimations
}
\description{
Calculate the second derivative with respect to the gamma estimations and phi estimations
}
