% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_FF_Rho}
\alias{calculateLL_FF_Rho}
\title{Calculate the second derivative with respect to the gamma estimations and rho estimations.}
\usage{
calculateLL_FF_Rho(B, N_i, p, k, phi, A, Dv, Dd, error_hat)
}
\arguments{
\item{B}{B matrix}

\item{N_i}{function matrices.}

\item{p}{autoregressive order of the error}

\item{k}{number of non parametric components}

\item{phi}{vector of the estimated phi parameters.}

\item{A}{A matrix}

\item{Dv}{diagonal of the  v1, ..., vn vectors}

\item{Dd}{diagonal of the d1, ..., dn vectors}

\item{error_hat}{residual of the model}
}
\value{
second derivative of the log function with respect to the gamma estimations and rho estimations
}
\description{
Calculate the second derivative with respect to the gamma estimations and rho estimations.
}
