% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_Phi}
\alias{calculateLL_Phi}
\title{Calculate the second derivative with respect to the phi estimations.}
\usage{
calculateLL_Phi(nn, phi, delta_i, Dc, Dv)
}
\arguments{
\item{nn}{number of observations.}

\item{phi}{vector of the estimated phi parameters.}

\item{delta_i}{delta_i}

\item{Dc}{diagonal of the  c1, ..., cn vectors}

\item{Dv}{diagonal of the  v1, ..., vn vectors}
}
\value{
second derivative of the log function with respect to the phi estimations
}
\description{
Calculate the second derivative with respect to the phi estimations.
}
