% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_Phi_Rho}
\alias{calculateLL_Phi_Rho}
\title{Calculate the second derivative with respect to the phi estimations and rho estimations}
\usage{
calculateLL_Phi_Rho(B, phi, error_hat, Dv, Dd, A)
}
\arguments{
\item{B}{B matrix}

\item{phi}{vector of the estimated phi parameters.}

\item{error_hat}{residual of the model}

\item{Dv}{diagonal of the v1, ..., vn vectors}

\item{Dd}{diagonal of the d1, ..., dn vectors}

\item{A}{A matrix}
}
\value{
second derivative of the log function with respect to the phi and rho estimations
}
\description{
Calculate the second derivative with respect to the phi estimations and rho estimations
}
