% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{calculateLL_Rho}
\alias{calculateLL_Rho}
\title{Calculate the second derivative with respect to the rho estimations.}
\usage{
calculateLL_Rho(B, p, phi, error_hat, Dd, Dv)
}
\arguments{
\item{B}{B matrix}

\item{p}{autoregressive order of the error}

\item{phi}{vector of the estimated phi parameters.}

\item{error_hat}{residual of the model}

\item{Dd}{diagonal of the d1, ..., dn vectors}

\item{Dv}{diagonal of the v1, ..., vn vectors}
}
\description{
Calculate the second derivative with respect to the rho estimations.
}
