% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.influence.aplms.R
\name{conf_normal_curvature}
\alias{conf_normal_curvature}
\title{Conformal normal curvature of local influence}
\usage{
conf_normal_curvature(DELTA, Lobs, Lobs.aux = 0)
}
\arguments{
\item{DELTA}{A numeric matrix, \eqn{\Delta}, that depends on the perturbation scheme.}

\item{Lobs}{A square numeric matrix representing the observed information matrix.}

\item{Lobs.aux}{An optional numeric matrix of the same dimension as
\code{Lobs}. Default is \code{0} (no auxiliary adjustment).}
}
\description{
Computes the conformal normal curvature.
}
