% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.aplms.R
\name{estimateVarF}
\alias{estimateVarF}
\title{Estimate the variance the asymptotic covariance-variance matrix for the estimated
gamma parameters.}
\usage{
estimateVarF(family, phi, const2, k, AN)
}
\arguments{
\item{family}{probability distribution function.}

\item{phi}{vector of the estimated phi parameters.}

\item{const2}{constant}

\item{k}{number of nn parametric components.}

\item{AN}{vector of the function matrices multiplied by the matrix A}
}
\description{
Estimate the variance the asymptotic covariance-variance matrix for the estimated
gamma parameters.
}
