% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{generateSummartTablePhiRho}
\alias{generateSummartTablePhiRho}
\title{Generate the summary table of the rho parameter.}
\usage{
generateSummartTablePhiRho(p, par1, rdf, rho, phi, nn, family)
}
\arguments{
\item{p}{autoregressive order of the error}

\item{par1}{optimized log-likelihood function}

\item{rdf}{regressive degrees of freedom}

\item{rho}{vector of estimated rho parameters}

\item{phi}{vector of estimated phi parameters}

\item{nn}{number of observations.}

\item{family}{probability distribution function.}
}
\description{
Generate the summary table of the rho parameter.
}
