% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{generateSummaryTable}
\alias{generateSummaryTable}
\title{Generate the summary table of the parametric components.}
\usage{
generateSummaryTable(model)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}
}
\description{
Generate the summary table of the parametric components.
}
