% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{generateSummaryTableRho}
\alias{generateSummaryTableRho}
\title{Generate the summary table of the rho parameter}
\usage{
generateSummaryTableRho(p, par1, rho, rdf)
}
\arguments{
\item{p}{autoregressive order of the error}

\item{par1}{optimized log-likelihood function}

\item{rho}{vector of estimated rho parameters}

\item{rdf}{regressive degrees of freedom}
}
\description{
Generate the summary table of the rho parameter
}
