% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{generateWaldF}
\alias{generateWaldF}
\title{Generate the Walf F statistic and its table}
\usage{
generateWaldF(npc_dimension, dfk, npc, f, VAR_F)
}
\arguments{
\item{npc_dimension}{cumulative sum of the number of rows of the function matrices}

\item{dfk}{effective degrees of freedon}

\item{npc}{vector with the names of non parametric components}

\item{f}{estimated gamma parameters}

\item{VAR_F}{covariance-variance matrix of the estimated gamma parameters}
}
\description{
Generate the Walf F statistic and its table
}
