% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.influence.aplms.R
\name{Gp}
\alias{Gp}
\title{Setting columns and rows indexes of a matrix to zero and compute the Generalized Inversion.}
\usage{
Gp(Lobs, index = NULL)
}
\arguments{
\item{Lobs}{The observed information matrix}

\item{index}{Indices of columns and rows to be set to zero}
}
\description{
Setting columns and rows indexes of a matrix to zero and compute the Generalized Inversion.
}
