% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitalization}
\alias{hospitalization}
\title{Respiratory diseases hospitalization Dataset}
\format{
The "data" slot is a data frame with 150 observations on the following 4 variables.
\describe{
\item{y}{dependent variable.}
\item{x1}{Independent variable 1.}
\item{x2}{Independent variable 2.}
}
}
\usage{
data(hospitalization)
}
\description{
This dataset consists of respiratory diseases hospitalization in Sorocaba, S\~ao Paulo, Brazil.
}
\examples{
data(hospitalization)
}
\references{
Include reference.
}
\keyword{datasets}
