% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.influence.aplms.R
\name{influence_DELTA}
\alias{influence_DELTA}
\title{Delta computation depending on the perturbation scheme}
\usage{
influence_DELTA(model, perturb_scheme = "case-weight", r = 1, k = 1)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{perturb_scheme}{A string vector specifying a perturbation scheme: `case-weight`, `dispersion`, `response`, `explanatory`, and `corAR`.}

\item{r}{Index of explanatory variable}

\item{k}{Index of autocorrelation coefficient}
}
\value{
A matrix Delta to support the perturbation scheme calculation.
}
\description{
Delta computation depending on the perturbation scheme
}
