% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.aplms.R
\name{pgam}
\alias{pgam}
\title{Piece of the P-GAM algorithm that creates the new estimation of the gamma parameters}
\usage{
pgam(A, N_i, Dv, phi, lam, K_i, f_init, y)
}
\arguments{
\item{A}{A matrix}

\item{N_i}{function matrices}

\item{Dv}{diagonal of the v1, ..., vn vectors}

\item{phi}{vector of the estimations of the phi parameters}

\item{lam}{smoothing parameter vector.}

\item{K_i}{function of matrices}

\item{f_init}{previous estimation of the gamma parameters}

\item{y}{vector of the observed values of the response variable}
}
\value{
The estimation of the gamma parameters
}
\description{
Piece of the P-GAM algorithm that creates the new estimation of the gamma parameters
}
