% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aplms.R
\name{plot.aplms}
\alias{plot.aplms}
\title{Default APLMS plotting.}
\usage{
\method{plot}{aplms}(model, level = 0.95, len = 100, plot = TRUE, ...)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{level}{Confidence level.}

\item{len}{The desired length of the sequence of covariates to compute the non parametric component functions.}

\item{plot}{a logical value to return plots. Default value is code{TRUE}.}

\item{...}{other arguments.}
}
\value{
Return a list of all non parametric component functions with their confidence intervals.
  interactive menu with eleven options to make plots.
}
\description{
Compute and plot the estimated mean and confidence intervals of the non-parametric component of a `APLMS` object fited by `aplms()`.
}
\keyword{Additive}
\keyword{Residuals}
\keyword{errors}
\keyword{linear}
\keyword{models}
\keyword{partial}
\keyword{symmetric}
\keyword{with}
