% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.aplms.R
\name{residuals.aplms}
\alias{residuals.aplms}
\title{Extract Residuals for APLMS fits}
\usage{
\method{residuals}{aplms}(object, ...)
}
\arguments{
\item{object}{an object with the result of fitting additive partial linear models with symmetric errors.
\code{response} indicates response residuals, \code{pearson} is Pearson residuals, and \code{quant} is quantile residuals.}

\item{...}{other arguments.}
}
\value{
a dataframe with the residual, the pearson residual and the normal quantile of the standarized resiudals.
}
\description{
Extract Residuals for APLMS fits
}
\examples{
\donttest{residuals(object)}
}
\keyword{Additive}
\keyword{Residuals}
\keyword{errors}
\keyword{linear}
\keyword{models}
\keyword{partial}
\keyword{symmetric}
\keyword{with}
