% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.influence.aplms.R
\name{set_zero_matrix}
\alias{set_zero_matrix}
\title{Setting columns and rows indexes of a matrix to zero}
\usage{
set_zero_matrix(mat, rows = NULL, cols = NULL)
}
\arguments{
\item{mat}{squared matrix}

\item{rows}{Row index}

\item{cols}{Column Index}
}
\description{
Setting columns and rows indexes of a matrix to zero
}
