% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{summary.aplms}
\alias{summary.aplms}
\alias{summary}
\title{Print method for "aplms" class}
\usage{
\method{summary}{aplms}(model, ...)

summary(model, ...)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "aplms" class
}
\examples{
\dontrun{summary(model)}
}
