% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aplms.R
\name{summary}
\alias{summary}
\alias{summary.aplms}
\title{Print method for "foo" class}
\usage{
summary(object, ...)

\method{summary}{aplms}(object, ...)
}
\arguments{
\item{object}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{...}{Other arguments passed to or from other methods}
}
\description{
Print method for "foo" class
}
