% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.aplms.R
\name{AIC.aplms}
\alias{AIC.aplms}
\title{Akaike information criterion}
\usage{
\method{AIC}{aplms}(model, ...)
}
\arguments{
\item{model}{APLMS object.}

\item{...}{other arguments}
}
\description{
Print the AIC of the fitted APLMS model.
}
\examples{
\dontrun{AIC(model)}
}
