% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplms.R
\name{aplms}
\alias{aplms}
\alias{print.aplms}
\title{Fitting Additive partial linear models with symmetric errors.}
\usage{
aplms(
  formula,
  npc,
  basis,
  Knot,
  data,
  family = Normal(),
  p = 1,
  control = list(tol = 0.001, algorithm1 = c("P-GAM"), algorithm2 = c("BFGS"), Maxiter1 =
    20, Maxiter2 = 25),
  init,
  lam
)

\method{print}{aplms}(model, ...)
}
\arguments{
\item{formula}{A symbolic description of the parametric component of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{npc}{A vector of names of non parametric component.}

\item{basis}{A vector of names of the basis to be used for each non parametric covariate.}

\item{Knot}{A vector of the number of knots in each non-linear component of the model.}

\item{data}{A data frame containing the variables in the model}

\item{family}{Symmetric error distribution}

\item{p}{autoregressive order of the error}

\item{control}{optimization rutine.}

\item{init}{A list of initial values for the symmetric error scale, phi, and autoregressive coefficients, rhos.}

\item{lam}{smoothing parameter vector.}

\item{model}{APLMS object.}

\item{...}{other arguments}
}
\description{
Fitting Additive partial linear models with symmetric errors
}
\details{
\code{aplms} is used to fit additive partial linear models with symmetric errors.
In this setup, the natural cubic splines or cubic P-splines.
}
\examples{
data(temperature)
datos = data.frame(temperature,time=1:length(temperature))
mod1<-aplms::aplms(temperature ~ 1,
                   npc=c("time"), basis=c("cr"),Knot=c(60),
                   data=datos,family=Powerexp(k=0.3),p=1,
                   control = list(tol = 0.001,
                                  algorithm1 = c("P-GAM"),
                                  algorithm2 = c("BFGS"),
                                  Maxiter1 = 20,
                                  Maxiter2 = 25),
                   lam=c(10))
summary(mod1)
}
