% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.aplms.R
\name{BIC.aplms}
\alias{BIC.aplms}
\title{Bayesian information criterion}
\usage{
\method{BIC}{aplms}(object, ...)
}
\arguments{
\item{object}{APLMS object.}

\item{...}{other arguments}
}
\value{
a numeric value of the corresponding BIC.
}
\description{
Print the BIC of the fitted APLMS model.
}
\examples{
\dontrun{BIC(object)}
}
