% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC.aplms.R
\name{BIC.aplms}
\alias{BIC.aplms}
\title{Bayesian information criterion}
\usage{
\method{BIC}{aplms}(model, ...)
}
\arguments{
\item{model}{APLMS object.}

\item{...}{other arguments}
}
\description{
Print the BIC of the fitted APLMS model.
}
\examples{
\dontrun{BIC(model)}
}
