% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.aplms.R
\name{coef.aplms}
\alias{coef.aplms}
\title{Extract the coefficients of the fitted APLMS model}
\usage{
\method{coef}{aplms}(model, ...)
}
\arguments{
\item{model}{APLMS object.}

\item{...}{other arguments}
}
\description{
Extract the coefficients of the fitted APLMS model.
}
\examples{
a <- 1
}
