% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.aplms.R
\name{coef.aplms}
\alias{coef.aplms}
\alias{coef}
\title{Extract the coefficients of the fitted APLMS model}
\usage{
\method{coef}{aplms}(model, ...)

coef(model, ...)
}
\arguments{
\item{model}{APLMS object.}

\item{...}{other arguments}
}
\value{
A list vector of the corresponding estimated parameters.
}
\description{
Extract the coefficients of the fitted APLMS model.
}
\examples{
\dontrun{coef(model)}
}
