% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.aplms.R
\name{influence.aplms}
\alias{influence.aplms}
\title{local influence analysis of the object `aplms()`.}
\usage{
\method{influence}{aplms}(
  model,
  perturbation = c("case-weight", "dispersion", "response", "explanatory", "corAR"),
  part = TRUE
)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{perturbation}{A string vector specifying a perturbation scheme: `case-weight`, `dispersion`, `response`, `explanatory`, and `corAR`.}

\item{part}{A logical value to indicate whether the influential analysis is performed for \eqn{\gamma}, \eqn{\phi} and \eqn{\rho}.}
}
\description{
Takes a fitted `aplms` object and outputs some diagnostic information about the fitting procedure and results. Returns the conformal normal curvature of the fitted `aplms` model object. The `case-weight`, `dispersion`, `response`, `explanatory`, and `corAR` perturbations are available.
}
\examples{
\dontrun{influence(model)}
}
