% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.aplms.R
\name{plot.aplms}
\alias{plot.aplms}
\title{Default APLMS plotting.}
\usage{
\method{plot}{aplms}(model, level = 0.95, len = 100, plot = TRUE, ...)
}
\arguments{
\item{model}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{level}{Confidence level.}

\item{len}{The desired length of the sequence of covariates to compute the non parametric component functions.}

\item{plot}{a logical value to return plots. Default value is code{TRUE}.}

\item{...}{other arguments.}
}
\value{
Return a list of all non parametric component functions with their confidence intervals.
  interactive menu with eleven options to make plots.
}
\description{
Compute and plot the estimated mean and confidence intervals of the non-parametric component of a `APLMS` object fited by `aplms()`.
}
\keyword{0.001,}
\keyword{1,}
\keyword{20,}
\keyword{25),}
\keyword{=}
\keyword{Additive}
\keyword{Maxiter1}
\keyword{Maxiter2}
\keyword{Residuals}
\keyword{algorithm1}
\keyword{algorithm2}
\keyword{basis=c("cr"),Knot=c(60),}
\keyword{c("BFGS"),}
\keyword{c("P-GAM"),}
\keyword{control}
\keyword{data(temperature)}
\keyword{data.frame(temperature,time=1:length(temperature))}
\keyword{data=datos,family=Powerexp(k=0.3),p=1,}
\keyword{datos}
\keyword{errors}
\keyword{lam=c(10))}
\keyword{linear}
\keyword{list(tol}
\keyword{mod1<-aplms::aplms(temperature}
\keyword{models}
\keyword{npc=c("time"),}
\keyword{partial}
\keyword{plot(mod1)}
\keyword{symmetric}
\keyword{with}
\keyword{~}
