% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.aplms.R
\name{residuals.aplms}
\alias{residuals.aplms}
\title{Extract Residuals for APLMS fits.}
\usage{
\method{residuals}{aplms}(mod, type = c("response", "pearson", "quant"), plot = NULL, ...)
}
\arguments{
\item{mod}{an object with the result of fitting additive partial linear models with symmetric errors.}

\item{type}{a character string that indicates the type of residuals.
\code{response} indicates response residuals, \code{pearson} is Pearson residuals, and \code{quant} is quantile residuals.}

\item{plot}{a logical evaluating.}

\item{...}{other arguments.}
}
\description{
Extract Residuals for APLMS fits.
}
\keyword{Additive}
\keyword{Residuals}
\keyword{errors}
\keyword{linear}
\keyword{models}
\keyword{partial}
\keyword{symmetric}
\keyword{with}
